added trees and minor changes

+++ src/main/java/com/herokuapp/player/Camera.java
@@ -1,12 +1,10 @@
 package com.herokuapp.player;
 
 public class Camera {
-	
-	public static double x = 0;
-	public static double y = 0;
-	
-	
-	
-	
+
+  public static int x = 0;
+  public static int y = 0;
+
+
 
 }
